#183 move unabstracted state into submodule
